finish adding actions to structurizer, placing is still a to-do

+++ Systems/Structurizer/PlacementActions/BaseRepeatedPlacementActionWithEntry.cs
@@ -5,7 +5,7 @@
     {
         public override ushort EntryData { get; }
 
-        protected BaseRepeatedPlacementActionWithEntry(byte repetitionCount, ushort entryData) : base(repetitionCount)
+        protected BaseRepeatedPlacementActionWithEntry(ushort repetitionCount, ushort entryData) : base(repetitionCount)
         {
             EntryData = entryData;
         }
